+++ packages/cozy-pouch-link/src/PouchManager.js
@@ -304,7 +304,10 @@ class PouchManager {
       localStorage.persistWarmedUpQueries(this.warmedUpQueries)
       logger.log('PouchManager: warmupQueries for ' + doctype + ' are done')
     } catch (err) {
-      logger.error('PouchManager: Could not warm up queries', err)
+      logger.error(
+        'PouchManager: Could not warm up queries for ' + doctype,
+        err
+      )
       delete this.warmedUpQueries[doctype]
     }
   }