+++ packages/cozy-pouch-link/CHANGELOG.md
@@ -3,6 +3,136 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [24.6.1](https://github.com/cozy/cozy-client/compare/v24.6.0...v24.6.1) (2021-09-13)
+
+**Note:** Version bump only for package cozy-pouch-link
+
+
+
+
+
+# [24.6.0](https://github.com/cozy/cozy-client/compare/v24.5.0...v24.6.0) (2021-09-13)
+
+**Note:** Version bump only for package cozy-pouch-link
+
+
+
+
+
+# [24.5.0](https://github.com/cozy/cozy-client/compare/v24.4.0...v24.5.0) (2021-09-09)
+
+**Note:** Version bump only for package cozy-pouch-link
+
+
+
+
+
+# [24.4.0](https://github.com/cozy/cozy-client/compare/v24.3.3...v24.4.0) (2021-09-08)
+
+**Note:** Version bump only for package cozy-pouch-link
+
+
+
+
+
+## [24.3.3](https://github.com/cozy/cozy-client/compare/v24.3.2...v24.3.3) (2021-09-07)
+
+**Note:** Version bump only for package cozy-pouch-link
+
+
+
+
+
+## [24.3.2](https://github.com/cozy/cozy-client/compare/v24.3.1...v24.3.2) (2021-09-03)
+
+
+### Bug Fixes
+
+* Method to get adapter name should be static ([ff7e26d](https://github.com/cozy/cozy-client/commit/ff7e26d40abc6c2a115df4cf3efcad5741242fc5))
+
+
+
+
+
+## [24.3.1](https://github.com/cozy/cozy-client/compare/v24.3.0...v24.3.1) (2021-09-03)
+
+
+### Bug Fixes
+
+* Correcty iterate over object keys ([9bf26cd](https://github.com/cozy/cozy-client/commit/9bf26cdd7f8ff7bd572cf98c79f116e0f9fbc8c6))
+
+
+
+
+
+# [24.3.0](https://github.com/cozy/cozy-client/compare/v24.2.0...v24.3.0) (2021-09-02)
+
+
+### Features
+
+* Expose method to get the adapter name ([2095d0a](https://github.com/cozy/cozy-client/commit/2095d0a4c0169bd10035043f2e6bd407dcae5bc3))
+
+
+
+
+
+# [24.2.0](https://github.com/cozy/cozy-client/compare/v24.1.3...v24.2.0) (2021-09-01)
+
+**Note:** Version bump only for package cozy-pouch-link
+
+
+
+
+
+## [24.1.3](https://github.com/cozy/cozy-client/compare/v24.1.2...v24.1.3) (2021-08-27)
+
+**Note:** Version bump only for package cozy-pouch-link
+
+
+
+
+
+## [24.1.2](https://github.com/cozy/cozy-client/compare/v24.1.1...v24.1.2) (2021-08-27)
+
+**Note:** Version bump only for package cozy-pouch-link
+
+
+
+
+
+## [24.1.1](https://github.com/cozy/cozy-client/compare/v24.1.0...v24.1.1) (2021-08-27)
+
+**Note:** Version bump only for package cozy-pouch-link
+
+
+
+
+
+# [24.1.0](https://github.com/cozy/cozy-client/compare/v24.0.0...v24.1.0) (2021-08-23)
+
+
+### Bug Fixes
+
+* Next attribute for Pouch ([556dd99](https://github.com/cozy/cozy-client/commit/556dd99fe0b66af993a9e08b6f658699cbda41b3))
+
+
+### Features
+
+* Add deleted true to a deleted doc with CozyPouchLink ([144be1d](https://github.com/cozy/cozy-client/commit/144be1d4293a1fdf1e4784fb9edbaebc3701cd21))
+* Add doctype specification in warmupQueries error ([b0f6d2f](https://github.com/cozy/cozy-client/commit/b0f6d2feb3a666a9f7e599d7a59d5b3c92f14fad))
+
+
+
+
+
+# [24.0.0](https://github.com/cozy/cozy-client/compare/v23.22.0...v24.0.0) (2021-08-13)
+
+**Note:** Version bump only for package cozy-pouch-link
+
+
+
+
+
 # [23.22.0](https://github.com/cozy/cozy-client/compare/v23.21.0...v23.22.0) (2021-08-12)